fixed setup logic

+++ main.py
@@ -2,14 +2,25 @@ import random
 import time
 import sys
 
+name = "default"
 def setup():
-    ans=input('do you want to login or register? ')
+    global name
+    ans=input('do you want to login or register?: ')
     if(ans=='register'):
         username = str(input('username: '))
         password = str(input('password: '))
         login = (username+password)
-        with open('usernames.txt', 'w') as f:
-            f.write(login)
+        confirm = str(input('this will overwrite any previous data, are you sure?(yes/no): '))
+        if(confirm=='yes'):
+            with open('usernames.txt', 'w') as f:
+                f.write(login)
+                name = login
+        elif(confirm=='no'):
+            print('returning to setup')
+            setup()
+        else:
+            print('unexpected input detected, returning to setup')
+            setup()
     elif(ans=='login'):
         username = str(input('username: '))
         password = str(input('password: '))
@@ -19,6 +30,7 @@ def setup():
                 if line == login:
                     print('you have logged in')
                     a = True
+                    name = line
                     break
                 else:
                     a = False
@@ -30,3 +42,4 @@ def setup():
         print('please either input "login" or "register"')
         setup()
 setup()
+print(name)
